refactor: move service configuration check logic to services module
- Add getUnconfiguredServiceIds() to TranslationServices class
- Simplify _filterUnconfiguredServices() in panel.ts to use new method
- Parse secrets JSON once instead of per-service for better performance
- Remove unused getPrefJSON import
